requirements.txt
# 化学信息学
rdkit-pypi>=2023.3.1b1

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 机器学习
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.4.0

# 深度学习
torch>=2.1.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torch-spline-conv>=1.2.0
transformers>=4.35.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# 实验管理
wandb>=0.16.0
mlflow>=2.8.0

# 其他工具
tqdm>=4.66.0
joblib>=1.3.0
